pyproject.toml
[project]
name = "nyc-taxi-analytics-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "apache-airflow==3.0.2",
    "apache-airflow-providers-google>=16.0.0",
    "apache-airflow-providers-snowflake>=6.4.0",
    "black>=25.1.0",
    "graphviz>=0.21",
    "snowflake-connector-python>=3.15.0",
]
